Bug 1752638 - Part 2: remote `waitpid`/`SIGCHLD` to/from the fork server. r=ipc-reviewers,nika
There are two parts to this:

1. A new sync IPC call to the fork server which is basically just
waitpid: it takes a pid and a flag for whether to block, and returns
either the exit status, an error, or an indication that the process is
still running.  The code in the parent process which would previously
hack around the fork server case with `kill(pid, 0)` now uses this
instead.

2. The fork server now has a `SIGCHLD` handler which writes to a pipe
given at startup.  Currently this is connected to the same pipe-to-self
used in `ProcessWatcher` for the parents' own `SIGCHLD`s.  Thus, the
existing `ProcessWatcher` is mostly unmodified.

Measured locally, these remote `waitpid`s take about 50µs, compared to
about 10µs for in-process.  So, while it's possible for the callback
for `SIGCHLD` pipe reads to call this `O(n)` times when `n` child
processes are exiting at once, the time taken shouldn't be a significant
responsiveness problem; I measured 250µs max while closing a tab for a
popular website with a large number of out-of-process iframes.  This can
be optimized later to batch the messages if needed.

The other potential performance issue is if a remote `waitpid` is blocked
behind a remote fork.  However, this is still an improvement over the
non-forkserver status quo: if we `clone` the parent process, that blocks
all threads for, empirically, about 15ms; cloning the fork server is
about 2.5ms, and that would only block another thread if it happens to
race with another child process exiting and not every time.  We also
prelaunch content processes during idle time, so this type of jank is
less likely to be user-visible in any case.

Importantly, the way this patch works, the `waitpid` is synchronous with
respect to the thread in the parent process making the requeste, so as
long as the fork server hasn't crashed then we can know that the process
with a given pid is in fact the process we expect it to be.

Differential Revision: https://phabricator.services.mozilla.com/D240714

UltraBlame original commit: 7c04f8c2c5a1f8d36f5915f753ece7195f66dc94

+++ ipc/chromium/src/chrome/common/process_watcher_posix_sigchld.cc
@@ -155,6 +155,18 @@ already_AddRefed<nsITimer> DelayedKill(pid_t aPid) {
   nsresult rv = NS_NewTimerWithCallback(
       getter_AddRefs(timer),
       [aPid](nsITimer*) {
+        
+        
+        
+        
+        
+        if (IsProcessDead(aPid, BlockingWait::No)) {
+          return;
+        }
+        
+        
+        
+        
         if (kill(aPid, SIGKILL) != 0) {
           const int err = errno;
       
@@ -171,6 +183,8 @@ already_AddRefed<nsITimer> DelayedKill(pid_t aPid) {
                                 << strerror(err);
           }
         }
+        
+        
       },
       kMaxWaitMs, nsITimer::TYPE_ONE_SHOT, "ProcessWatcher::DelayedKill",
       XRE_GetAsyncIOEventTarget());
@@ -382,7 +396,21 @@ static void ProcessWatcherInit() {
       }));
 }
 
+static void EnsureProcessWatcher() {
+  static std::once_flag sInited;
+  std::call_once(sInited, ProcessWatcherInit);
+}
+
 }  
+
+mozilla::UniqueFileHandle ProcessWatcher::GetSignalPipe() {
+  EnsureProcessWatcher();
+  int fd = gSignalPipe[1];
+  MOZ_ASSERT(fd >= 0);
+  fd = dup(fd);
+  MOZ_ASSERT(fd >= 0);
+  return mozilla::UniqueFileHandle(fd);
+}
 
 
 
@@ -406,8 +434,7 @@ void ProcessWatcher::EnsureProcessTerminated(base::ProcessHandle process,
   DCHECK(process != base::GetCurrentProcId());
   DCHECK(process > 0);
 
-  static std::once_flag sInited;
-  std::call_once(sInited, ProcessWatcherInit);
+  EnsureProcessWatcher();
 
   auto lock = gPendingChildren.Lock();
   auto& children = lock.ref();
@@ -442,12 +469,14 @@ void ProcessWatcher::EnsureProcessTerminated(base::ProcessHandle process,
         
         
         
+        
+        
+        
+        
 
         CHROMIUM_LOG(WARNING) << "EnsureProcessTerminated: duplicate process"
                                  " ID "
-                              << process
-                              << "; assuming this is because of the fork"
-                                 " server.";
+                              << process;
